bugfix: fix referer implementation (#3766)
* chore: empty commit for PR creation

* chore: test with correct URL

* chore: add referer to wagmi

* chore: alter fallback provider

* chore: add same config for wagmi client

* chore: add logging

* chore: try removing staticNetwork

* chore: try using a fallback for clientToSigner

* chore: update clientToSigner

* chore: update clientToSigner per wagmi docs

* chore: test header manually

* chore: test headers manually in transports

* chore: add clean implementation on ethers/transports

* chore: rm logging

* chore: add NEXT_PUBLIC_APP_ENVIRONMENT to nextjs build test github workflow

* bugfix: add staticNetwork functionality back

* chore: add NEXT_PUBLIC_APP_ENVIRONMENT env and test referal

* chore: rm test referral URL

---------
